...crucial use of @export to make the functions appear in the namespace
I worked out why library(socialFrontier) still didn't load the functions.

To export a function to namespace we use @export. Else the function can only be accessed using pkg::: methods. Packages not in the namespace are hidden from the user -- not intended to be used.

Finally we have to install and rebuild.

Major:
1) used @export in the royxgen2 tags to export functiosn to namespace so they can actually be used