preAllocation tested
The routine still uses two parts -- 1) a part using st_intersect in a for loop and 2) another part combining the results of the st_intersect into one sf object

This method uses preallocation in step two and preallocates empty rows in a dataset to be filled.

The new function works and prints timing for both stages. We find that for barnet the prellocate

Results: preAllocate doesn't seem to be faster

Major:
1) Created the preAllocate code
2) Added results in the test script

Minor:
1) code for printing times for both parts