rbindlist from data.table [Tested]
For putting the list together we use a rbindlist from data.table.

Weakness of this approach is that every item in the list has to be the same so we need it all to be the same sf geometry throughout.

Code is changed to st_cast as multiline after st_instersect

The routine is now very fast.

Major:

1) added option rbindlist from data.table as a method.
2) Added st_cast after st_intersect to turn all instersections into lines.

Minor:
1) ignoring R history files

Issue:
st_cast may have deprcated code at the end of the function that extract all line intersections. This is more applicable to earlier methods.