Check of rbindlist for frontier_as_sf
Check and built package to see if using rbindlist instead of do.call(rbind) works as intended.

Everything is fine and documentation is build. rbindlist is now the default method.

Only issue is that the bound box is wrong so for plotting (using default tmap behaviour) causes issues under very specific cases. This is udpated in the test script

Major:
1) Tested rbind list
2) produced code replicating the tmap plot issue with the method
3) updated documentation to indicate the issue and default uses rbindlist